pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper-dic"
dynamic = ["version"]
requires-python = ">=3.12"
description = "Voice-to-text dictation tool with hotkey activation"
dependencies = [
    "pynput>=1.8.0",
    "sounddevice>=0.5.0",
    "soundfile>=0.12.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "pyperclip>=1.9.0",
]

[project.optional-dependencies]
macos = [
    "pyobjc-framework-Quartz>=10.0",
    "rumps>=0.4.0",
]

[project.scripts]
whisper-dic = "whisper_dic.cli:main"

[tool.setuptools.dynamic]
version = {file = "src/whisper_dic/VERSION"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
whisper_dic = ["VERSION", "config.example.toml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_incomplete_defs = true
warn_return_any = true
warn_unused_configs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "Quartz",
    "Quartz.*",
    "AppKit",
    "AppKit.*",
    "PyObjCTools",
    "PyObjCTools.*",
    "ApplicationServices",
    "ApplicationServices.*",
    "rumps",
    "rumps.*",
    "pynput",
    "pynput.*",
    "sounddevice",
    "sounddevice.*",
    "soundfile",
    "soundfile.*",
    "pyperclip",
    "pyperclip.*",
    "pychromecast",
    "pychromecast.*",
    "async_upnp_client",
    "async_upnp_client.*",
    "tomllib",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "whisper_dic.compat._windows"
ignore_errors = true
